clang-tidy: fix too small loop variable
Found with bugprone-too-small-loop-variable

+++ src/tags_int.cpp
@@ -2170,7 +2170,7 @@ namespace Exiv2 {
         {
             uint16_t bit   = 0;
             uint16_t comma = 0;
-            for (uint16_t i = 0; i < value.count(); i++ ) { // for each element in value array
+            for (long i = 0; i < value.count(); i++) {  // for each element in value array
                 auto bits = static_cast<uint16_t>(value.toLong(i));
                 for (uint16_t b = 0; b < 16; ++b) { // for every bit
                     if (bits & (1 << b)) {
